refactor(llm): move MockLLMProvider from production code to test doubles module
MockLLMProvider was defined in app/core/llm/providers.py alongside
production providers, with no environment guard. Move it to
tests/llm_test_doubles.py where test doubles belong. The class is not
imported by any test file and is not registered in the provider factory.

Closes #193
